requirements.txt
aiohappyeyeballs==2.6.1
PyYAML>=6.0
watchdog>=4.0
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
frozenlist==1.8.0
gkeepapi==0.16.0
idna==3.11
keyring==25.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
Pygments==2.19.2
requests==2.32.5
rich==14.3.3
samsungtvws==3.0.4
shellingham==1.5.4
typer==0.24.1
urllib3==2.6.3
wakeonlan==3.1.0
websocket-client==1.9.0
websockets==16.0
yarl==1.23.0
